+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.1] - 2019-03-28
 ### Added
 - Add `rustfmt` to CI and move in that direction.
 - Add coordinator version to `replictl versions`.
@@ -87,5 +89,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - WebUI endpoints for initial UI.
 
 
-[Unreleased]: https://github.com/replicante-io/replicante/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/replicante-io/replicante/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/replicante-io/replicante/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/replicante-io/replicante/compare/v0.1.0...v0.2.0